create initial stock status component skeleton, add 'order by' display function
this commit creates the initial file and structure for the Stock Status and 'order by' section
of the description. This section tells the user if the item is in stock, and if it is
tells them when to order by to receive their item within two days. This functionality
will depend on the stock status attribute stored in the database, which in turn is
store in Apps state.